EventEndpoint erweitert

+++ SkatenightBackend/src/main/java/ws1415/SkatenightBackend/EventEndpoint.java
@@ -3,16 +3,14 @@ package ws1415.SkatenightBackend;
 import com.google.api.server.spi.config.Named;
 import com.google.appengine.api.oauth.OAuthRequestException;
 import com.google.appengine.api.users.User;
-import com.googlecode.objectify.Key;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Date;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
-import java.util.logging.Logger;
 
 import javax.jdo.PersistenceManager;
 import javax.jdo.Query;
@@ -21,13 +19,15 @@ import ws1415.SkatenightBackend.gcm.Message;
 import ws1415.SkatenightBackend.gcm.MessageType;
 import ws1415.SkatenightBackend.gcm.RegistrationManager;
 import ws1415.SkatenightBackend.gcm.Sender;
-import ws1415.SkatenightBackend.model.BooleanWrapper;
-import ws1415.SkatenightBackend.model.Event;
-import ws1415.SkatenightBackend.transport.EventMetaData;
-import ws1415.SkatenightBackend.model.Member;
-import ws1415.SkatenightBackend.model.Route;
 import ws1415.SkatenightBackend.model.EndUser;
+import ws1415.SkatenightBackend.model.Event;
+import ws1415.SkatenightBackend.model.EventRole;
+import ws1415.SkatenightBackend.model.Member;
+import ws1415.SkatenightBackend.model.Privilege;
+import ws1415.SkatenightBackend.model.Route;
 import ws1415.SkatenightBackend.model.UserLocation;
+import ws1415.SkatenightBackend.transport.EventMetaData;
+import ws1415.SkatenightBackend.transport.EventParticipationData;
 
 import static com.googlecode.objectify.ObjectifyService.ofy;
 
@@ -36,6 +36,42 @@ import static com.googlecode.objectify.ObjectifyService.ofy;
  * @author Richard Schulze
  */
 public class EventEndpoint extends SkatenightServerEndpoint {
+
+    /**
+     * Weist dem Benutzer mit der angegebenen Email
+     * @param user       Der aufrufende Benutzer.
+     * @param eventId    Die ID des Events, in dem die Rolle zugewiesen werden soll.
+     * @param endUser    Der Benutzer, dem die Rolle zugewiesen werden soll.
+     * @param role       Die zuzuweisende Rolle.
+     */
+    public void assignRole(User user, @Named("eventId") long eventId, @Named("endUser") String endUser, @Named("role") EventRole role) throws OAuthRequestException {
+        if (user == null) {
+            throw new OAuthRequestException("no user submitted");
+        }
+        Event event = ofy().load().group(EventParticipationData.class).type(Event.class).id(eventId).safe();
+        if (hasPrivilege(event, user.getEmail(), Privilege.ASSIGN_ROLE)) {
+            event.getMemberList().put(endUser, role);
+            ofy().save().entity(event).now();
+        } else {
+            throw new OAuthRequestException("insufficient privileges");
+        }
+    }
+
+    /**
+     * Prüft ob der angegebene Benutzer in dem übergebenen Event das entsprechende Privileg besitzt.
+     * @param event     Das Event, für das das Privileg geprüft werden soll.
+     * @param endUser   Die Mail-Adresse des zu prüfenden Benutzers.
+     * @param privilege Das zu prüfende Privileg.
+     * @return true, wenn der Benutzer das Privileg im angegebenen Event besitzt, sonst false.
+     */
+    private boolean hasPrivilege(Event event, String endUser, Privilege privilege) {
+        // TODO Methode ggf. mit User-Objekt sichern
+        if (event == null || endUser == null || privilege == null
+                || !event.getMemberList().containsKey(endUser)) {
+            return false;
+        }
+        return event.getMemberList().get(endUser).hasPrivilege(privilege);
+    }
 
     /**
      * Gibt eine Liste aller auf dem Server gespeicherter Metadaten von Events zurück.
@@ -71,13 +107,30 @@ public class EventEndpoint extends SkatenightServerEndpoint {
      * @return Das persistierte Event-Objekt.
      */
     // TODO Route hinzufügen
-    public Event createEvent(User user, Event event) {
-        // TODO User prüfen
+    public Event createEvent(User user, Event event) throws OAuthRequestException {
+        if (user == null) {
+            throw new OAuthRequestException("no user submitted");
+        }
+        // TODO Globale Rolle des aufrufenden Benutzers prüfen
+        //if (new RoleEndpoint().getGlobalRole(user.getEmail()).value != Role.HOST.getId()) {
+        //    throw new OAuthRequestException("user is not a host");
+        //}
         // TODO Event prüfen
+
+        // Den aufrufenden Benutzer als Host eintragen
+        if (event.getMemberList() == null) {
+            event.setMemberList(new HashMap<String, EventRole>());
+        }
+        event.getMemberList().put(user.getEmail(), EventRole.HOST);
 
         ofy().save().entity(event).now();
 
         return event;
+    }
+
+
+    public void editEvent(User user, Event event) {
+        // TODO
     }
 
     /**
@@ -86,12 +139,56 @@ public class EventEndpoint extends SkatenightServerEndpoint {
      * @param user Der aufrufende Benutzer.
      * @param id   Die ID des zu löschenden Events.
      */
-    public void deleteEvent(User user, @Named("id") Long id) {
+    public void deleteEvent(User user, @Named("id") Long id) throws OAuthRequestException {
         if (id != null) {
+            if (user == null) {
+                throw new OAuthRequestException("no user submitted");
+            }
             // TODO User prüfen
             // TODO Gallery und alle zugehörigen Blobs löschen
 
             ofy().delete().type(Event.class).id(id);
+        }
+    }
+
+    /**
+     * Fügt den aufrufenden Teilnehmer zum angegebenen Event hinzu.
+     * @param user       Der aufrufende Benutzer, der dem Event hinzugefügt werden soll
+     * @param eventId    Die ID des Events, dem beigetreten wird.
+     */
+    public void joinEvent(User user, @Named("eventId") Long eventId) throws OAuthRequestException {
+        if (user == null) {
+            throw new OAuthRequestException("no user submitted");
+        }
+        if (eventId != null) {
+            Event event = ofy().load().group(EventParticipationData.class).type(Event.class).id(eventId).safe();
+            if (!event.getMemberList().containsKey(user.getEmail())) {
+                event.getMemberList().put(user.getEmail(), EventRole.PARTICIPANT);
+                new UserEndpoint().getFullUser(user.getEmail()).addEvent(event);
+                // TODO Muss geändertes EndUser-Objekt gespeichert werden?
+            }
+        }
+    }
+
+    /**
+     * Entfernt den aufrufenden Benutzer aus dem Event mit der angegebenen ID.
+     * @param user       Der zu entfernende User.
+     * @param eventId    Die ID des Events, das verlassen wird.
+     * @throws IllegalStateException falls der letzte Host versucht aus dem Event auszutreten.
+     */
+    public void leaveEvent(User user, @Named("eventId") Long eventId) throws OAuthRequestException {
+        if (user == null) {
+            throw new OAuthRequestException("no user submitted");
+        }
+        if (eventId != null) {
+            Event event = ofy().load().group(EventParticipationData.class).type(Event.class).id(eventId).safe();
+            event.getMemberList().remove(user.getEmail());
+            if (!event.getMemberList().containsValue(EventRole.HOST)) {
+                throw new IllegalStateException("the last host of an event can not leave");
+            }
+            new UserEndpoint().getFullUser(user.getEmail()).removeEvent(event);
+            // TODO Muss geändertes EndUser-Objekt gespeichert werden?
+            ofy().save().entity(event).now();
         }
     }
 
@@ -222,10 +319,9 @@ public class EventEndpoint extends SkatenightServerEndpoint {
      */
     public void addMemberToEvent(@Named("id") long keyId, @Named("email") String email) {
         Event event = getEvent(keyId);
-        ArrayList<String> memberKeys = event.getMemberList();
+        Set<String> memberKeys = event.getMemberList().keySet();
         if (!memberKeys.contains(email)) {
-            memberKeys.add(email);
-            event.setMemberList(memberKeys);
+            event.getMemberList().put(email, EventRole.PARTICIPANT);
 
             updateEvent(event);
 
@@ -271,11 +367,9 @@ public class EventEndpoint extends SkatenightServerEndpoint {
     public void removeMemberFromEvent(@Named("id") long keyId, @Named("email") String email) {
         Event event = getEvent(keyId);
 
-        ArrayList<String> memberKeys = event.getMemberList();
+        Set<String> memberKeys = event.getMemberList().keySet();
         if (memberKeys.contains(email)) {
-            memberKeys.remove(email);
-            event.setMemberList(memberKeys);
-
+            event.getMemberList().remove(email);
             updateEvent(event);
         }
     }
@@ -290,7 +384,7 @@ public class EventEndpoint extends SkatenightServerEndpoint {
         Event event = getEvent(keyId);
 
         List<EndUser> endUsers = new ArrayList<>(event.getMemberList().size());
-        for (String email: event.getMemberList()) {
+        for (String email: event.getMemberList().keySet()) {
             endUsers.add(new UserEndpoint().getFullUser(email));
         }
         return endUsers;
@@ -306,7 +400,7 @@ public class EventEndpoint extends SkatenightServerEndpoint {
         Event event = getEvent(keyId);
 
         List<UserLocation> userLocations = new ArrayList<>();
-        for (String email: event.getMemberList()) {
+        for (String email: event.getMemberList().keySet()) {
             userLocations.add(new UserEndpoint().getUserLocation(email));
         }
         return userLocations;
@@ -337,25 +431,6 @@ public class EventEndpoint extends SkatenightServerEndpoint {
         finally {
             pm.close();
         }
-    }
-
-    /**
-     * Bearbeitet das Event mit der Id des übergebenen Events mit den Daten des übergebenen Events.
-     *
-     * @param event das zu bearbeitende Event mit den neuen Daten
-     * @param user User, der die Methode aufruft
-     * @return true, wenn Aktion erfolgreich, false sonst
-     * @throws OAuthRequestException
-     * @throws IOException
-     */
-    public BooleanWrapper editEvent(Event event, User user) throws OAuthRequestException, IOException {
-        long keyId = event.getId();
-        Date oldDate = getEvent(keyId).getDate();
-        Date newDate = event.getDate();
-        boolean dateChanged = oldDate.equals(newDate);
-        deleteEvent(user, event.getId());
-        createEvent(user, event, true, dateChanged);
-        return new BooleanWrapper(true);
     }
 
     /**
